Updated RPCDDS installation process on CentOS 6.

+++ utils/installers/rti/linux/setup_linux_rpcdds.sh
@@ -168,9 +168,11 @@ function installer
 
 function rpminstaller
 {
-    rm tmp/$project/classes/antlr-2.7.7.jar
-    rm tmp/$project/classes/stringtemplate-3.2.1.jar
-    rm tmp/$project/classes/rpcddsgen.jar
+        rm tmp/$project/classes/antlr-2.7.7.jar
+        rm tmp/$project/classes/rpcddsgen.jar
+        if [[ "${distroversion}" == "Fedora"* ]]; then 
+            rm tmp/$project/classes/stringtemplate-3.2.1.jar
+        fi
 
 	# Change the script form local to general script.
 	cp ../../../../scripts/rpcddsgen_rpm.sh tmp/$project/scripts/rpcddsgen.sh
@@ -182,7 +184,7 @@ function rpminstaller
 	mkdir -p tmp/$project/fastrpcgen
 
 	# Copy the build.xml
-	if [ ${distroversion} == CentOS6.4 ]; then
+	if [[ "${distroversion}" == "CentOS6"* ]]; then
 		cp build_rpm_rpcdds_centos.xml tmp/$project/fastrpcgen/build.xml
 	else
 		cp build_rpm_rpcdds.xml tmp/$project/fastrpcgen/build.xml
@@ -220,8 +222,8 @@ function rpminstaller
 	if [ $errorstatus != 0 ]; then return; fi
 
 	# Copy SPEC file
-	if [ ${distroversion} == CentOS6.4 ]; then
-		sed "s/VERSION/${version}/g" FastBuffers_centos.spec > ~/rpmbuild/SPECS/FastBuffers.spec
+	if [[ "${distroversion}" == "CentOS6"* ]]; then
+		sed "s/VERSION/${version}/g" rpcdds_centos.spec > ~/rpmbuild/SPECS/rpcdds.spec
 	else
 		sed "s/VERSION/${version}/g" rpcdds.spec > ~/rpmbuild/SPECS/rpcdds.spec
 	fi
@@ -247,7 +249,11 @@ function rpminstaller
 
     # Install fastcdr for i686
     cd ../RPMS/i686
-    sudo yum localinstall fastcdr-0.2.2-1.fc20.i686.rpm
+    if [[ "${distroversion}" = "Fedora20" ]]; then
+        sudo yum localinstall fastcdr-0.2.2-1.fc20.i686.rpm
+    elif [[ "${distroversion}" = "CentOS6"* ]]; then
+        sudo yum localinstall fastcdr-0.2.2-1.el6.i686.rpm
+    fi
     cd -
 
 	# Build command for i686.
@@ -260,7 +266,11 @@ function rpminstaller
 
     # Install fastcdr for x64
     cd ../RPMS/x86_64
-    sudo yum localinstall fastcdr-0.2.2-1.fc20.x86_64.rpm
+    if [ "${distroversion}" = "Fedora20" ]; then
+        sudo yum localinstall fastcdr-0.2.2-1.fc20.x86_64.rpm
+    elif [[ "${distroversion}" = "CentOS6"* ]]; then
+        sudo yum localinstall fastcdr-0.2.2-1.el6.x86_64.rpm
+    fi
     cd -
 
 	# Build command for x86_64.
@@ -290,8 +300,11 @@ mkdir tmp/$project
 
 installer
 
-# TODO Detect if the distro suport RPM
-[ $errorstatus == 0 ] && { rpminstaller; }
+if [ $errorstatus == 0 ]; then
+    if [[ "${distroversion}" == "CentOS6"* ]] || [[ "${distroversion}" == "Fedora"* ]]; then
+        rpminstaller
+    fi
+fi
 
 # Remove temporaly directory
 rm -rf tmp
